Add shutdown hook
