+++ src/main/scala/wvlet/airframe/LifeCycleEventHandler.scala
@@ -65,57 +65,3 @@ class LifeCycleEventHandlerPair(parent:LifeCycleEventHandler, child:LifeCycleEve
   }
 }
 
-object ShowLifeCycleLog extends LifeCycleEventHandler {
-
-  private val logger = Logger.of[LifeCycleManager]
-
-  override def beforeStart(lifeCycleManager:LifeCycleManager) {
-    logger.info(s"Life cycle is starting ...")
-  }
-
-  override def afterStart(lifeCycleManager:LifeCycleManager) {
-    logger.info(s"======= STARTED =======")
-  }
-
-  override def beforeShutdown(lifeCycleManager:LifeCycleManager) {
-    logger.info(s"Stopping life cycle ...")
-  }
-
-  override def afterShutdown(lifeCycleManager:LifeCycleManager) {
-    logger.info(s"Life cycle has stopped.")
-  }
-}
-
-
-object JSR330AnnotationHandler extends LifeCycleEventHandler with LogSupport {
-  override def onInit(lifeCycleManager:LifeCycleManager, t: ObjectType, injectee: AnyRef) {
-    debug(s"Injected ${t} (class: ${t.rawType}): $injectee")
-    val schema = ObjectSchema(t.rawType)
-
-    // Find JSR330 @PostConstruct annotation
-    schema
-    .allMethods
-    .filter{_.findAnnotationOf[PostConstruct].isDefined}
-    .map{x =>
-      lifeCycleManager.addInitHook(ObjectMethodCall(injectee, x))
-    }
-
-    // Find JSR330 @PreDestroy annotation
-    schema
-    .allMethods
-    .filter {_.findAnnotationOf[PreDestroy].isDefined}
-    .map { x =>
-      lifeCycleManager.addShutdownHook(ObjectMethodCall(injectee, x))
-    }
-  }
-}
-
-object FIFOHookExecutor extends LifeCycleEventHandler {
-  override def beforeStart(lifeCycleManager: LifeCycleManager): Unit = {
-    lifeCycleManager.startHooks.reverse.map(_.execute)
-  }
-
-  override def beforeShutdown(lifeCycleManager: LifeCycleManager): Unit = {
-    lifeCycleManager.shutdownHooks.map(_.execute)
-  }
-}
